Move what can be add common for mod_lbmethod_cluster.
