+++ native/balancers/mod_lbmethod_cluster.c
@@ -20,6 +20,23 @@
 #include "apr_version.h"
 #include "ap_hooks.h"
 
+#include "slotmem.h"
+
+#include "node.h"
+#include "host.h"
+#include "context.h"
+#include "balancer.h"
+
+#include "mod_proxy_cluster.h"
+
+static struct node_storage_method *node_storage = NULL;
+static struct host_storage_method *host_storage = NULL;
+static struct context_storage_method *context_storage = NULL;
+static struct balancer_storage_method *balancer_storage = NULL;
+static struct domain_storage_method *domain_storage = NULL;
+
+static int use_alias = 0; /* 1 : Compare Alias with server_name */
+
 module AP_MODULE_DECLARE_DATA lbmethod_cluster_module;
 
 static proxy_worker *find_best(proxy_balancer *balancer,
@@ -63,39 +80,98 @@ static const proxy_balancer_method cluster =
  */
 static int lbmethod_cluster_trans(request_rec *r)
 {
-    int i;
+    const char *balancer;
     void *sconf = r->server->module_config;
     proxy_server_conf *conf = (proxy_server_conf *)
         ap_get_module_config(sconf, &proxy_module);
-    char *ptr = conf->balancers->elts;
-    int sizeb = conf->balancers->elt_size;
-    char *balancername = NULL;
 
 
+#if HAVE_CLUSTER_EX_DEBUG
     ap_log_error(APLOG_MARK, APLOG_NOERRNO|APLOG_DEBUG, 0, r->server,
                 "lbmethod_cluster_trans for %d %s %s uri: %s args: %s unparsed_uri: %s",
                  r->proxyreq, r->filename, r->handler, r->uri, r->args, r->unparsed_uri);
     ap_log_error(APLOG_MARK, APLOG_NOERRNO|APLOG_DEBUG, 0, r->server,
                 "lbmethod_cluster_trans for %d", conf->balancers->nelts);
-    for (i = 0; i < conf->balancers->nelts; i++, ptr=ptr+sizeb) {
-        proxy_balancer *balancer = (proxy_balancer *) ptr;
-        ap_log_error(APLOG_MARK, APLOG_NOERRNO|APLOG_DEBUG, 0, r->server,
-                    "lbmethod_cluster_trans for %s", balancer->s->name);
-        balancername = balancer->s->name;
-    }
-    if (!balancername)
-       return DECLINED;
+#endif
 
-    if (strncmp(r->uri, "/examples", 9) == 0) {
-        r->filename =  apr_pstrcat(r->pool, "proxy:", balancername, r->uri, NULL);
+    proxy_vhost_table *vhost_table = read_vhost_table(r, host_storage);
+    proxy_context_table *context_table = read_context_table(r, context_storage);
+    proxy_balancer_table *balancer_table = read_balancer_table(r, balancer_storage);
+    proxy_node_table *node_table = read_node_table(r, node_storage);
+
+    apr_table_setn(r->notes, "vhost-table",  (char *) vhost_table);
+    apr_table_setn(r->notes, "context-table",  (char *) context_table);
+    apr_table_setn(r->notes, "balancer-table",  (char *) balancer_table);
+    apr_table_setn(r->notes, "node-table",  (char *) node_table);
+
+    balancer = get_route_balancer(r, conf, vhost_table, context_table, balancer_table, node_table, use_alias);
+    if (!balancer) {
+        balancer = get_context_host_balancer(r, vhost_table, context_table, node_table, use_alias);
+    }
+    
+
+    if (balancer) {
+
+        /* It is safer to use r->uri */
+        if (strncmp(r->uri, "balancer://",11))
+            r->filename =  apr_pstrcat(r->pool, "proxy:balancer://", balancer, r->uri, NULL);
+        else
+            r->filename =  apr_pstrcat(r->pool, "proxy:", r->uri, NULL);
         r->handler = "proxy-server";
         r->proxyreq = PROXYREQ_REVERSE;
-    ap_log_error(APLOG_MARK, APLOG_NOERRNO|APLOG_DEBUG, 0, r->server,
-                "lbmethod_cluster_trans (OK) for %d %s %s uri: %s args: %s unparsed_uri: %s",
-                 r->proxyreq, r->filename, r->handler, r->uri, r->args, r->unparsed_uri);
-        return OK;
+#if HAVE_CLUSTER_EX_DEBUG
+        ap_log_error(APLOG_MARK, APLOG_NOERRNO|APLOG_DEBUG, 0, r->server,
+                    "proxy_cluster_trans using %s uri: %s",
+                     balancer, r->filename);
+#endif
+        return OK; /* Mod_proxy will process it */
     }
+ 
+#if HAVE_CLUSTER_EX_DEBUG
+    ap_log_error(APLOG_MARK, APLOG_NOERRNO|APLOG_DEBUG, 0, r->server,
+                "proxy_cluster_trans DECLINED %s uri: %s unparsed_uri: %s",
+                 balancer, r->filename, r->unparsed_uri);
+#endif
     return DECLINED;
+}
+
+static int lbmethod_cluster_post_config(apr_pool_t *p, apr_pool_t *plog,
+                                     apr_pool_t *ptemp, server_rec *s)
+{
+   node_storage = ap_lookup_provider("manager" , "shared", "0");
+    if (node_storage == NULL) {
+        ap_log_error(APLOG_MARK, APLOG_ERR|APLOG_NOERRNO, 0, s,
+                    "proxy_cluster_post_config: Can't find mod_manager for nodes");
+        return !OK;
+    }
+    host_storage = ap_lookup_provider("manager" , "shared", "1");
+    if (host_storage == NULL) {
+        ap_log_error(APLOG_MARK, APLOG_ERR|APLOG_NOERRNO, 0, s,
+                    "proxy_cluster_post_config: Can't find mod_manager for hosts");
+        return !OK;
+    }
+    context_storage = ap_lookup_provider("manager" , "shared", "2");
+    if (context_storage == NULL) {
+        ap_log_error(APLOG_MARK, APLOG_ERR|APLOG_NOERRNO, 0, s,
+                    "proxy_cluster_post_config: Can't find mod_manager for contexts");
+        return !OK;
+    }
+    balancer_storage = ap_lookup_provider("manager" , "shared", "3");
+    if (balancer_storage == NULL) {
+        ap_log_error(APLOG_MARK, APLOG_ERR|APLOG_NOERRNO, 0, s,
+                    "proxy_cluster_post_config: Can't find mod_manager for balancers");
+        return !OK;
+    }
+    domain_storage = ap_lookup_provider("manager" , "shared", "5");
+    if (domain_storage == NULL) {
+        ap_log_error(APLOG_MARK, APLOG_ERR|APLOG_NOERRNO, 0, s,
+                    "proxy_cluster_post_config: Can't find mod_manager for domains");
+        return !OK;
+    }
+
+    /* Add version information */
+    ap_add_version_component(p, MOD_CLUSTER_EXPOSED_VERSION);
+    return OK;
 }
 
 static void register_hooks(apr_pool_t *p)
@@ -106,6 +182,7 @@ static void register_hooks(apr_pool_t *p)
     ap_register_provider(p, PROXY_LBMETHOD, "cluster", "0", &cluster);
 
     ap_hook_translate_name(lbmethod_cluster_trans, aszPre, aszSucc, APR_HOOK_FIRST);
+    ap_hook_post_config(lbmethod_cluster_post_config, NULL, NULL, APR_HOOK_MIDDLE);
 }
