+++ native/mod_proxy_cluster/Makefile.in
@@ -16,8 +16,6 @@ top_builddir = @APACHE_BUILDDIR@
 # For .deps.
 builddir = @CLUSTER_BASE@
 
-MOD_OBJS_LO=
-
 include $(top_builddir)/build/rules.mk
 SH_COMPILE = $(LIBTOOL) --mode=compile $(BASE_CC) -I../include -prefer-pic -c $< && touch $@
 
@@ -26,9 +24,9 @@ all: mod_proxy_cluster.so
 mod_proxy_cluster.so: mod_proxy_cluster.la
 	$(top_builddir)/build/instdso.sh SH_LIBTOOL='$(LIBTOOL)' mod_proxy_cluster.la `pwd`
 
-mod_proxy_cluster.la: mod_proxy_cluster.slo
-	$(SH_LINK) -rpath $(libexecdir) -module -avoid-version mod_proxy_cluster.lo $(MOD_OBJS_LO)
+mod_proxy_cluster.la: mod_proxy_cluster.slo ../common/common.slo
+	$(SH_LINK) -rpath $(libexecdir) -module -avoid-version mod_proxy_cluster.lo common.lo
 
 clean:
-	rm -f *.o *.lo *.slo *.so
+	rm -f *.o *.lo *.slo *.so ../common/common.slo
 	rm -rf .libs
